Moved the creation class
I had the creation class in the wrong folder so i moved it to generators
