+++ lib/enum_weasel.rb
@@ -1,7 +1,7 @@
 require 'active_record'
 require "enum_weasel/version"
 require "enum_weasel/configuration"
-require "enum_weasel/tables/creation"
+require "enum_weasel/generator/tables/creation"
 require "enum_weasel/get_model_enums"
 require "enum_weasel/generator"
